use only one DeleteDialogActivity
