+++ OpenKeychain/src/main/AndroidManifest.xml
@@ -686,9 +686,6 @@
             android:name=".ui.DeleteKeyDialogActivity"
             android:theme="@android:style/Theme.NoDisplay" />
         <activity
-            android:name=".ui.RevokeDeleteDialogActivity"
-            android:theme="@android:style/Theme.NoDisplay" />
-        <activity
             android:name=".ui.OrbotRequiredDialogActivity"
             android:theme="@android:style/Theme.NoDisplay" />
         <activity android:name=".ui.PassphraseWizardActivity" />
